✨ feat: bookmark 쿼리 뼈대 개발
- 서버 통합 후 추가 개발 예정
