+++ src/shared/api/getMyBookmarks.ts
@@ -1,0 +1,2 @@
+const getMyBookmarks = (userId: number) => {};
+export default getMyBookmarks;
