+++ src/widgets/popular-plans/api/getPopularPlans.ts
@@ -1,0 +1,2 @@
+const getPopularPlans = () => {};
+export default getPopularPlans;
