+++ src/widgets/popular-plans/mock/data.ts
@@ -1,19 +1,4 @@
-'use client';
-
-import { useCarousel } from '@/shared';
-import {
-  MdOutlineKeyboardArrowLeft,
-  MdOutlineKeyboardArrowRight,
-} from 'react-icons/md';
-import usePopularPlan from '../model/usePopularPlan';
-import PopularPlan from './PopularPlan';
-
-const BREAK_POINTS: { [key: number]: number } = {
-  320: 1,
-  574: 2,
-  1024: 4,
-};
-const data = [
+export const data = [
   {
     plan_id: 1,
     user_id: 1,
@@ -271,54 +256,3 @@ const data = [
     schedules: [],
   },
 ];
-
-export default function PopularPlans() {
-  //  const { data, isFetching, isLoading } = usePopularPlan();
-  const { carouselStartIndex, visibleSlides, nextItem, prevItem } = useCarousel(
-    BREAK_POINTS,
-    data ? data.length : 0,
-    false,
-    0
-  );
-  return (
-    <div className="space-y-4">
-      <div>
-        <h2 className="text-xl font-bold">금주 인기 여행 일정</h2>
-      </div>
-      <div className="relative flex w-full justify-center overflow-hidden px-8 py-4">
-        {data && (
-          <ul className=" flex w-full justify-between overflow-hidden">
-            <button
-              type="button"
-              className="absolute left-0 top-2/4 z-10 -translate-y-2/4  text-4xl disabled:hidden"
-              aria-label="이전 슬라이드"
-              onClick={prevItem}
-              disabled={carouselStartIndex === 0}
-            >
-              <MdOutlineKeyboardArrowLeft />
-            </button>
-            {data?.map((plan) => (
-              <PopularPlan
-                key={plan.plan_id}
-                plan={plan}
-                carouselStartIndex={carouselStartIndex}
-                visibleSlides={visibleSlides}
-              />
-            ))}
-            <button
-              className="absolute right-0 top-1/2 -translate-y-1/2 text-4xl disabled:hidden"
-              type="button"
-              aria-label="다음 슬라이드"
-              onClick={nextItem}
-              disabled={
-                carouselStartIndex >= (data?.length ?? 0) - visibleSlides
-              }
-            >
-              <MdOutlineKeyboardArrowRight />
-            </button>
-          </ul>
-        )}
-      </div>
-    </div>
-  );
-}
