+++ src/widgets/popular-plans/model/usePopularPlans.ts
@@ -1,0 +1,7 @@
+import { useQuery } from '@tanstack/react-query';
+import getPopularPlans from '../api/getPopularPlans';
+
+const usePopularPlans = () => {
+  return useQuery({ queryKey: ['popularPlans'], queryFn: getPopularPlans });
+};
+export default usePopularPlans;
