--- src/widgets/popular-schedule/model/usePopularPlan.ts
@@ -1,6 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-
-const usePopularPlan = () => {
-  return useQuery;
-};
-export default usePopularPlan;
